Fix the loop cutter and add a dataflow limit.
The problem is, even in the absence of loops, the dataflow sometimes is processed multiple times on
the same edge. It doesn't know that the graph is a DAG, so it doesn't traverse it in a reasonable
order, so in general every vertex could be processed dozens of times. It's prohibitively slow.